rag_project/New folder/backend_requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-chroma==0.1.0
langchain-community==0.0.10
langgraph==0.0.10
openai==1.3.0
chromadb==0.4.14
pypdf==3.17.1
